asm-verify: fix ulong/int width + const signature mismatches (#266 Cluster D)
Restore binary signatures so the mangled names pair (unsigned long = m,
unsigned int = j, int = i -- distinct even where all 4 bytes on the port):
- Font::GetPage(unsigned long) const -- was int
- Skeleton::BuildArrays(unsigned long) -- was int
- Touch::FindTouch/GetTouchPos/GetTouchDelta -- param was already uint;
  real divergence was const-ness (binary is non-const) -> drop const
- DojoScreen::UpdateBSButton(BSButton*, float, unsigned long) -- 3rd was int
- Bonus::IsAchieved map key std::map<unsigned long,int> -- was uint32_t
  (j); changed m_MinFruit/m_MaxFruit/m_RequiredHashes keys + IsAchieved/
  GetBest/UnlockAchievements iterators consistently. Behavior-safe:
  unsigned long is 4 bytes on ARM32/Windows LLP64, sizeof(std::map)
  unchanged, offset static_asserts still hold, hashes are 32-bit anyway.

Mangling-only on the port (LLP64), but fixes asm-verify pairing on the
Bada cross-build. Call sites pass int literals -> implicit convert, no
call-site changes. triage.json was partly stale (Touch param + Bonus key
were not what it claimed) -- verified each vs current source first.

Verified: fresh build + 79/79 logic ctest.

+++ src/engine/asset/Skeleton.cpp
@@ -63,8 +63,8 @@ uint32_t Skeleton::FindIndex(const char* name) const {
 // Binary allocates ONE block new Matrix44[count*3], partitions into 3 pointers.
 // No per-element Identity() -- memory left uninitialized.
 // Port uses resize (still calls Matrix44 ctor per element; port limitation).
-void Skeleton::BuildArrays(int count) {
-    if (count == (int)m_LocalMatrices.size() && !m_LocalMatrices.empty()) return;
+void Skeleton::BuildArrays(unsigned long count) {
+    if (count == m_LocalMatrices.size() && !m_LocalMatrices.empty()) return;
     if (count == 0) {
         m_LocalMatrices.clear();
         m_WorldMatrices.clear();
